+++ src/main/resources/templates/product/cart.html
@@ -20,7 +20,7 @@
         <div class="order-button">
             <button onclick="location.href='/orders'"><img src="https://cdn-icons-png.flaticon.com/512/3144/3144456.png"
                                                            alt="주문 목록"/></button>
-            <button onclick="location.href='/products/list'">
+            <button onclick="location.href='/admin/products/list'">
                 <img src="./assets/product_list_icon.png" alt="상품 목록"/></button>
         </div>
     </div>
