+++ src/main/resources/templates/product/list.html
@@ -2,39 +2,40 @@
 <html lang="ko" xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta charset="UTF-8">
-    <title>상품 목록</title>
+    <title>관리자 - 상품 목록</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
 </head>
 <body>
 <div class="container mt-5">
-    <h1 class="mb-4">상품 목록</h1>
+    <h1 class="mb-4">관리자 - 상품 목록</h1>
     <div class="d-flex gap-2 mb-3">
-        <a th:href="@{/products/new}" class="btn btn-primary">새 상품 추가</a>
+        <a th:href="@{/admin/products/new}" class="btn btn-primary">새 상품 추가</a>
         <a th:href="@{/products}" class="btn btn-secondary">장바구니로 돌아가기</a>
     </div>
 
-    <table class="table table-striped">
-        <thead>
-        <tr>
-            <th>ID</th>
-            <th>이름</th>
-            <th>가격</th>
-            <th>이미지</th>
-            <th>작업</th>
-        </tr>
-        </thead>
-        <tbody>
-        <tr th:each="product : ${productResponseDtos}">
-            <td th:text="${product.productId}"></td>
-            <td th:text="${product.name}"></td>
-            <td th:text="${product.price}"></td>
+    <div class="table-responsive">
+        <table class="table table-striped">
+            <thead>
+            <tr>
+                <th>ID</th>
+                <th>이름</th>
+                <th>가격</th>
+                <th>이미지</th>
+                <th>작업</th>
+            </tr>
+            </thead>
+            <tbody>
+            <tr th:each="product : ${productResponseDtos}">
+                <td th:text="${product.productId}"></td>
+                <td th:text="${product.name}"></td>
+                <td th:text="${product.price}"></td>
+                <td>
+                    <img th:if="${product.imageUrl}" th:src="${product.imageUrl}" alt="상품 이미지" width="50"/>
+                    <span th:unless="${product.imageUrl}">이미지 없음</span>
+                </td>
             <td>
-                <img th:if="${product.imageUrl}" th:src="${product.imageUrl}" alt="상품 이미지" width="50"/>
-                <span th:unless="${product.imageUrl}">이미지 없음</span>
-            </td>
-            <td>
-                <a th:href="@{/products/{id}/edit(id=${product.productId})}" class="btn btn-sm btn-warning">수정</a>
-                <form th:action="@{/products/{id}(id=${product.productId})}" method="post" style="display:inline;">
+                <a th:href="@{/admin/products/{id}/edit(id=${product.productId})}" class="btn btn-sm btn-warning">수정</a>
+                <form th:action="@{/admin/products/{id}(id=${product.productId})}" method="post" style="display:inline;">
                     <input type="hidden" name="_method" value="DELETE">
                     <button type="submit" class="btn btn-sm btn-danger" onclick="return confirm('정말 삭제하시겠습니까?')">삭제</button>
                 </form>
